splitted poller.cpp up in files
